feat: Move DTOS to their own infrastructure folder for a more coherent structure
