+++ src/Chirp.Infrastructure/DataTransferObjects/FollowDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace Chirp.Core;
+﻿namespace Chirp.Infrastructure.DataTransferObjects;
 
 public class FollowDto
 {
