+++ src/Chirp.Web/Areas/Identity/Pages/Account/Manage/Index.cshtml.cs
@@ -5,6 +5,7 @@
 
 using Chirp.Core;
 using Chirp.Infrastructure.Chirp.Services;
+using Chirp.Infrastructure.DataTransferObjects;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
