+++ src/Chirp.Web/Pages/Shared/TimelineModel.cs
@@ -1,5 +1,6 @@
 ﻿using Chirp.Core;
 using Chirp.Infrastructure.Chirp.Services;
+using Chirp.Infrastructure.DataTransferObjects;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Build.Framework;
