+++ src/Chirp.Web/Program.cs
@@ -50,13 +50,11 @@ using ( var serviceScope = app.Services.CreateScope() )
 }
 
 
-//DbInitializer.SeedDatabase( app.Services.GetService<CheepDBContext>() );
-
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    // The default HSTS value is 30 days. 
     app.UseHsts();
 }
